Dishes search pr addition

+++ src/pages/Menus/MenuPage.jsx
@@ -10,6 +10,8 @@ const MenuPage = () => {
   const location = useLocation();
   const highlightDishId = location.state?.highlightDishId;
   const { menuId } = useParams();
+  const [error, setError] = useState('');
+  const [loading, setLoading] = useState(false);
   const [selectedDish, setSelectedDish] = useState(null);
   const [dishes, setDishes] = useState([]);
   const [newDishId, setNewDishId] = useState([]);
@@ -24,21 +26,50 @@ const MenuPage = () => {
       try {
         const data = await dishService.getMenuByid(menuId);
         setDishes(data.dishes);
-      } catch (err) {
-        console.error("Greška pri učitavanju jela:", err);
+      } catch (error) {
+        if (error.response) {
+          if (error.response.status === 404) {
+            setError("Menu sa ovim id-em ne postoji ili pogresna ruta.");
+          } else if (error.response.status === 500) {
+            setError("Greska na serveru. Pokusajte kasnije.");
+          } else {
+            setError(`Greska: ${error.response.status}`);
+          }
+        } else if (error.request) {
+          setError("Nema odgovora sa servera.");
+        } else {
+          setError("Doslo je do greske.");
+        }
+        console.error("Greška pri slanju zahteva za permisije:", error);
+      } finally {
+        setLoading(false);
       }
     };
     const fetchRestaurant = async () => {
-      const permit = JSON.parse(sessionStorage.getItem('permitRequest'));
-      if (!permit || permit != true) {
-        return;
-      }
+      const permit = JSON.parse(sessionStorage.getItem("permitRequest"));
+      if (!permit || permit != true) return;
       try {
         const response = await getMenuPermissionAsync(menuId);
-        sessionStorage.removeItem('permitRequest');
         setIsOwnerHere(response == true);
-      } catch (err) {
-        console.error("Greška pri učitavanju restorana:", err);
+      } catch (error) {
+        if (error.response) {
+          if (error.response.status === 404) {
+            setError("Jelo sa ovim id-em ne postoji ili pogresna ruta.");
+          } else if (error.response.status === 401) {
+            setError("Ova ruta je rezervisana samo za vlasnike restorana.");
+          } else if (error.response.status === 500) {
+            setError("Greska na serveru. Pokusajte kasnije.");
+          } else {
+            setError(`Greska: ${error.response.status}`);
+          }
+        } else if (error.request) {
+          setError("Nema odgovora sa servera.");
+        } else {
+          setError("Doslo je do greske.");
+        }
+        console.error("Greška pri slanju zahteva za permisije:", error);
+      } finally {
+        setLoading(false);
       }
     };
     fetchMenu();
@@ -52,21 +83,50 @@ const MenuPage = () => {
   }, {});
 
   const handleSave = async (dish) => {
+    const formData = new FormData();
+    formData.append("Name", dish.name)
+    formData.append("Type", dish.type)
+    formData.append("Price", dish.price)
+    formData.append("Description", dish.description)
+    formData.append("MenuId", menuId)
+    if (dish.picture) {
+      formData.append("file", dish.picture[0]);
+    }
     try {
       if (dish.id) {
-      await dishService.update(dish.id, dish);
+      formData.append("Id", dish.id)
+      await dishService.update(dish.id, formData);
 
       const data = await dishService.getAll();
       setDishes(data);
       } else {
-        const created = await dishService.create(dish);
+        const created = await dishService.create(formData);
         setDishes((prev) => [...prev, created]);
 
         setNewDishId(created.id);
         setIsDishGroupOpen(true);
       }
-    } catch (err) {
-      console.error("Greška pri snimanju:", err);
+    } catch (error) {
+      if (error.response) {
+        if (error.response.status === 400) {
+          setError('Niste uneli validne podatke.');
+        } else if (error.response.status === 404) {
+          setError("Jelo sa ovim id-em ne postoji ili pogresna ruta.");
+        } else if (error.response.status === 401) {
+          setError("Ova ruta je rezervisana samo za vlasnike restorana.");
+        } else if (error.response.status === 500) {
+          setError("Greska na serveru. Pokusajte kasnije.");
+        } else {
+          setError(`Greska: ${error.response.status}`);
+        }
+      } else if (error.request) {
+        setError("Nema odgovora sa servera.");
+      } else {
+        setError("Doslo je do greske.");
+      }
+      console.error("Greska:", error.message);
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -76,14 +136,43 @@ const MenuPage = () => {
   }
 
   const deleteDish = async (id) => {
+    if (!window.confirm("Da li ste sigurni da zelite da uklonite jelo?")) {
+      return;
+    }
     try {
+      setLoading(true);
       await dishService.delete(id);
       setDishes((prev) => prev.filter((d) => d.id !== id));
-    } catch (err) {
-      console.error("Greška pri brisanju jela:", err);
+      setError("");
+      alert("Uspesno ste uklonili jelo");
+    } catch (error) {
+      if (error.response) {
+        if (error.response.status === 404) {
+          setError("Ne postoji jelo sa ovim id-em.");
+        } else if (error.response.status === 500) {
+          setError("Greska na serveru. Pokusajte kasnije.");
+        } else {
+          setError(`Greska: ${error.response.status}`);
+        }
+      } else if (error.request) {
+        setError("Nema odgovora sa servera.");
+      } else {
+        setError("Doslo je do greske.");
+      }
+      console.error("Greska:", error.message);
+    } finally {
+      setLoading(false);
     }
   };
 
+  useEffect(() => {
+    return () => {
+      sessionStorage.removeItem('permitRequest');
+    };
+  }, []);
+
+  if (loading) return <div id="loadingSpinner" className="spinner"></div>;
+  if (error) return <p style={{ color: 'red' }}>{error}</p>;
   return (
     <div style={{ padding: "20px" }}>
       <h1 style={{ fontSize: "28px", marginBottom: "0", fontStyle: 'italic' }}>Glavni Menu</h1>
